requirements-gpu.txt
# apt-get update && apt-get install ffmpeg libsm6 libxext6  -y

--extra-index-url https://download.pytorch.org/whl/cu118


# facechain

Pillow
opencv-python
modelscope==1.10.0

# accelerate
transformers
onnxruntime==1.15.1
# diffusers>=0.23.0
# invisible-watermark>=0.2.0
# modelscope==1.10.0

torchvision==0.16

# mmengine
# numpy==1.22.0
# protobuf==3.20.1
# timm
# scikit-image==0.19.3
# gradio
# controlnet_aux==0.0.6
# mediapipe
# python-slugify
mmdet==2.26.0
SentencePiece
# # mmcv-full (need mim install)
openmim

# # https://mmdetection.readthedocs.io/en/v2.26.0/get_started.html
# # pip install -U openmim -i https://pypi.tuna.tsinghua.edu.cn/simple
# # mim install mmcv-full


# pip install -U openmim -i https://pypi.tuna.tsinghua.edu.cn/simple
# mim install mmcv  
# /usr/local/cuda/bin/nvcc -V
# export CUDA_HOME=/usr/local/cuda-11.8
# cp /usr/local/cuda/bin/nvcc /usr/bin/

